requirements.txt
flask==2.3.3
tensorflow==2.14.0
pillow==9.5.0
numpy==1.24.3
scikit-learn==1.3.0